requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.4
django-crispy-forms==1.14.0
factory-boy==3.2.1
Faker==13.7.0
idna==3.3
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
